requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
Django==4.1.7
django-daraja==1.2.0
django-environ==0.10.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.4
phonenumbers==8.13.7
pycparser==2.21
python-decouple==3.8
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
whitenoise==6.4.0
